+++ SpatialGDK/Source/SpatialGDKFunctionalTests/SpatialGDK/SpatialTestPossession/SpatialTestRemotePossession.h
@@ -21,12 +21,14 @@ struct FControllerPawnPair
 	APawn* Pawn = nullptr;
 };
 
-UCLASS()
+UCLASS(Abstract)
 class SPATIALGDKFUNCTIONALTESTS_API ASpatialTestRemotePossession : public ASpatialFunctionalTest
 {
 	GENERATED_BODY()
 public:
 	ASpatialTestRemotePossession();
+
+	ASpatialTestRemotePossession(const EMapCategory MapCiCategory, const int32 NumberOfClients);
 
 	virtual void PrepareTest() override;